Blackstlink: Stlinkv2 handling in its own file.
