+++ src/platforms/blacklink/Makefile.inc
@@ -7,4 +7,4 @@ CFLAGS += -Wno-cast-function-type
 else ifneq (, $(findstring cygwin, $(SYS)))
 LDFLAGS += -lws2_32
 endif
-SRC += 	timing.c	\
+SRC += 	timing.c stlinkv2.c
